add date arg to clearAppGroupContainer to only delete files older than date

+++ ios/ExpoShareExtensionModule.swift
@@ -13,7 +13,7 @@ public class ExpoShareExtensionModule: Module {
       NotificationCenter.default.post(name: NSNotification.Name("openHostApp"), object: nil, userInfo: userInfo)
     }
 
-    AsyncFunction("clearAppGroupContainer") { (promise: Promise) in
+    AsyncFunction("clearAppGroupContainer") { (date: String?, promise: Promise) in
       DispatchQueue.global(qos: .background).async {
         guard let appGroup = Bundle.main.object(forInfoDictionaryKey: "AppGroup") as? String else {
           DispatchQueue.main.async {
@@ -29,6 +29,21 @@ public class ExpoShareExtensionModule: Module {
           return
         }
 
+        var comparisonDate: Date? = nil
+        if let isoDateString = date {
+          let isoFormatter = ISO8601DateFormatter()
+          // new Date().toISOString() returns fractional seconds, which we need to account for:
+          isoFormatter.formatOptions.insert(.withFractionalSeconds)
+          if let date = isoFormatter.date(from: isoDateString) {
+            comparisonDate = date
+          } else {
+            DispatchQueue.main.async {
+              promise.reject("ERR_INVALID_DATE", "The provided date string is not in a valid ISO 8601 format")
+            }
+            return
+          }
+        }
+
         let fileManager = FileManager.default
         let sharedDataUrl = containerUrl.deletingLastPathComponent().appendingPathComponent("sharedData")
 
@@ -37,7 +52,20 @@ public class ExpoShareExtensionModule: Module {
             let contents = try fileManager.contentsOfDirectory(atPath: sharedDataUrl.path)
             for item in contents {
               let itemPath = sharedDataUrl.appendingPathComponent(item).path
-              try fileManager.removeItem(atPath: itemPath)
+              if let creationDate = self.getCreationDate(of: itemPath) {
+                if let comparisonDate = comparisonDate {
+                  if creationDate < comparisonDate {
+                    try fileManager.removeItem(atPath: itemPath)
+                  }
+                } else {
+                  try fileManager.removeItem(atPath: itemPath)
+                }
+              } else {
+                DispatchQueue.main.async {
+                  promise.reject("ERR_REMOVE_CONTENTS", "Unable to retrieve creation date")
+                }
+                return
+              }
             }
             DispatchQueue.main.async {
               print("sharedData directory contents removed successfully.")
@@ -56,5 +84,18 @@ public class ExpoShareExtensionModule: Module {
         }
       }
     }
+  }
+
+  internal func getCreationDate(of filePath: String) -> Date? {
+    let fileManager = FileManager.default
+    do {
+      let attributes = try fileManager.attributesOfItem(atPath: filePath)
+      if let creationDate = attributes[.creationDate] as? Date {
+        return creationDate
+      }
+    } catch {
+      print("Error getting file attributes: \(error.localizedDescription)")
+    }
+    return nil
   }
 }
